refactor(cli): extract the serialize.log ledger subsystem into ledger.py
Pure move: the ledger writers, parse regexes, session-ledger fold,
outstanding computation, and heal plan relocate verbatim from cli.py
(2173 -> 1916 lines) into ledger.py. Every moved segment is a byte-
identical substring of the previous cli.py, comments included, and
status/heal output was captured before and after on a fixture ledger:
checksums equal.

cli.py re-exports all eighteen moved names so existing cli.<name>
references (tests, in-flight branches) resolve unchanged. The stats
fold still lives in cli.py and now uses the imported regexes — moving
it is a natural later slice, kept out of this one to stay a pure move.
